+++ modules/apps/staging/staging-processes-web/src/main/java/com/liferay/staging/processes/web/portlet/action/ProcessesListMVCRenderCommand.java
@@ -12,15 +12,10 @@
  * details.
  */
 
-package com.liferay.staging.processes.web.portlet.action.process;
+package com.liferay.staging.processes.web.portlet.action;
 
-import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCResourceCommand;
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.staging.processes.web.constants.StagingProcessesPortletKeys;
-
-import javax.portlet.PortletRequestDispatcher;
-import javax.portlet.ResourceRequest;
-import javax.portlet.ResourceResponse;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -31,22 +26,15 @@ import org.osgi.service.component.annotations.Component;
 	immediate = true,
 	property = {
 		"javax.portlet.name=" + StagingProcessesPortletKeys.STAGING_PROCESSES,
-		"mvc.command.name=publishLayouts"
+		"mvc.command.name=processesList"
 	},
-	service = MVCResourceCommand.class
+	service = MVCRenderCommand.class
 )
-public class PublishLayoutsMVCResourceCommand extends BaseMVCResourceCommand {
+public class ProcessesListMVCRenderCommand extends GetGroupMVCRenderCommand {
 
 	@Override
-	protected void doServeResource(
-			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
-		throws Exception {
-
-		PortletRequestDispatcher portletRequestDispatcher =
-			getPortletRequestDispatcher(
-				resourceRequest, "/processes_list/view.jsp");
-
-		portletRequestDispatcher.include(resourceRequest, resourceResponse);
+	protected String getPath() {
+		return "/view.jsp";
 	}
 
 }
