+++ modules/apps/staging/staging-processes-web/src/main/java/com/liferay/staging/processes/web/portlet/action/PublishLayoutsMVCRenderCommand.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.staging.processes.web.portlet.action.publication;
+package com.liferay.staging.processes.web.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.staging.processes.web.constants.StagingProcessesPortletKeys;
 
+import com.liferay.staging.processes.web.portlet.action.GetGroupMVCRenderCommand;
 import org.osgi.service.component.annotations.Component;
 
 /**
